models/team28_BVISR/requirements.txt
transformers>=4.28.1
einops>=0.7.0
open-clip-torch>=2.20.0
peft>=0.9.0
Pillow>=9.5.0
PyYAML>=6.0
pyiqa>=0.1.10
huggingface_hub>=0.25.0
numpy<2.0.0
opencv-python>=4.11.0.86
basicsr
loralib
fairscale
lpips